+++ packages/cli/src/extensions/remote/connection.ts
@@ -28,6 +28,7 @@ import { emitSessionActive } from "./chunked-delivery.js";
 import { resetRelayRegistrationGate, signalRelayRegistered } from "./registration-gate.js";
 import { decideRegisteredParentState } from "../remote-registered-parent-state.js";
 import { waitForWorkerStartupComplete } from "../worker-startup-gate.js";
+import { resolveInputDeliverAs } from "./deliver-as-default.js";
 
 // ── Module-level singletons (safe: one relay extension per process) ───────────
 
@@ -75,32 +76,6 @@ export interface ConnectionHandlers {
     onSocketTeardown: () => void;
     /** Compute parentSessionId for register (accounts for pendingDelinkOwnParent). */
     getParentSessionIdForRegister: () => string | null | undefined;
-}
-
-// ── deliverAs defaulting ──────────────────────────────────────────────────────
-
-/**
- * Resolve the effective `deliverAs` for a user input message.
- *
- * When the web UI sends a message while it believes the agent is idle, it
- * intentionally omits `deliverAs`. If the runtime turns out to be streaming
- * by the time the message is ready to dispatch (typical after a slow MCP
- * startup where the initial prompt has already started streaming), passing
- * `undefined` into `prompt()` would throw "Agent is already processing..."
- * and the message would be silently dropped.
- *
- * If no explicit mode was supplied and the agent is currently streaming,
- * default to `"followUp"` so the message is safely queued.
- *
- * Exported for unit testing; consumed by the `input` socket handler.
- */
-export function resolveInputDeliverAs(
-    requested: "followUp" | "steer" | undefined,
-    isAgentActive: boolean,
-): "followUp" | "steer" | undefined {
-    if (requested) return requested;
-    if (isAgentActive) return "followUp";
-    return undefined;
 }
 
 // ── URL helpers ───────────────────────────────────────────────────────────────
